refactor(agent-compile): consolidate scripts; design_space on the full split
Structural cleanup of scripts/agent_compile/ (the experiment-space redesign,
skill prune, and graph_expand removal already landed via #196/#190; this is the
scripts dedup that did not):

- scripts/agent_compile/lib/{config,harness,prebuilt}.py: shared, importable —
  no script imports a sibling. SweepConfig supports a base+overlay YAML and an
  empty 'instances' meaning the full split.
- One runner (run_sweep.py, absorbs run_agent_sweep.py; legacy run_sweep deleted)
  and one aggregator (aggregate.py, absorbs aggregate_phase2.py — per-arm
  metrics, invocation histogram, easy/hard, per-scenario, Pareto; drops the
  CAR-only compile_table derivation). run_sweep validates the harness (raises on
  unknown tool/skill ids) before spending.
- Offline retrieval ablations -> scripts/retrieval_ablation/; compare_harnesses
  -> scripts/analysis/. Agent-localization scorer upstreamed to
  codeminer/eval/retrieval_eval.py:score_agent_localization (one definition).
- configs: drop the fragmented cost_*/phase0/sample (different instance slices,
  not co-plottable); keep design_space.yaml as the canonical multi-arm config,
  extended to the FULL codeminer-base test split so all 9 arms are comparable.
- partition.py + build_codeminer_base_partition.py (#190 fit/heldout artifact)
  kept untouched. Docs (README + design ADR + experiments) reframed.

Verified on origin/main base: flake8 clean; design_space loads (9 arms, full
split) + passes the harness guard; test_aggregate + test_prebuilt green; full
suite collects (1107). Runner/aggregator mechanics validated end-to-end by a
12-cell smoke (all arm types incl. read-only + no-fs).
