--- scripts/agent_compile/configs/cost_context.yaml
@@ -1,31 +1,0 @@
-# SPDX-FileCopyrightText: 2025-2026 CodeMiner Contributors
-#
-# SPDX-License-Identifier: Apache-2.0
-# Graph-aware harness: defaults (grep/read) ON + the one-call codeminer_context
-# composer (search -> graph-expand -> compact assembly). Tests whether a single
-# deterministic-graph call reduces the grep/read fan-out.
-sweep_id: cost_context
-model: vertex_ai/claude-haiku-4-5
-vertex_location: us-east5
-vertex_project: null
-reps: 1
-max_turns: 16
-temperature: 0.0
-max_tokens: 4096
-topk: 50
-num_retries: 8
-dataset: fishmingyu/codeminer-base-dataset
-split: test
-prebuilt_dir: /mnt/data/codeminer
-embedding_model: Qwen/Qwen3-Embedding-0.6B
-embedding_dimension: 1024
-metrics_k: [1, 3, 5, 10]
-gt_simplified_symbols: true
-include_default_tools: true
-instances:
-  - babel__babel-15445
-  - vuejs__core-11589
-  - micropython__micropython-13569
-subsets:
-  CTX:
-    - codeminer_context
