--- scripts/agent_compile/configs/cost_grep.yaml
@@ -1,43 +1,0 @@
-# SPDX-FileCopyrightText: 2025-2026 CodeMiner Contributors
-#
-# SPDX-License-Identifier: Apache-2.0
-# Pure grep/read baseline: ONLY the default tool layer (read/grep/glob/bash).
-# The subset names the default TOOL ids: tools are not skills, so this matches
-# ZERO retrieval skills (the arm exposes only the always-on default tools) while
-# staying non-empty (an empty subset would trip AgentRunner's empty->full-
-# registry fallback — see AgentRunner._resolve_allow_set). This is "the grep/read
-# agent" the graph-aware composer must beat on tokens at equal accuracy.
-# (Equivalent to design_space.yaml's defaults_only arm.)
-sweep_id: cost_grep
-model: vertex_ai/claude-haiku-4-5
-vertex_location: us-east5
-vertex_project: null
-reps: 1
-max_turns: 16
-temperature: 0.0
-max_tokens: 4096
-topk: 50
-num_retries: 8
-dataset: fishmingyu/codeminer-base-dataset
-split: test
-prebuilt_dir: /mnt/data/codeminer
-embedding_model: Qwen/Qwen3-Embedding-0.6B
-embedding_dimension: 1024
-metrics_k: [1, 3, 5, 10]
-gt_simplified_symbols: true
-include_default_tools: true
-instances:
-  - redis__redis-10095
-  - tokio-rs__tokio-4898
-  - facebook__docusaurus-10130
-  - axios__axios-4731
-  - hashicorp__terraform-34814
-  - caddyserver__caddy-5870
-  - pydata__xarray-2905
-  - sympy__sympy-13031
-subsets:
-  GREP:
-    - read
-    - grep
-    - glob
-    - bash
