--- scripts/agent_compile/configs/cost_locagent.yaml
@@ -1,70 +1,0 @@
-# SPDX-FileCopyrightText: 2025-2026 CodeMiner Contributors
-#
-# SPDX-License-Identifier: Apache-2.0
-# LocAgent-style GRAPH-PRIMARY harness: the agent gets bm25 + call-graph tools
-# and file_read to confirm, but NO grep (file_search withheld via
-# default_tool_ids). Tests whether removing the grep escape hatch makes the
-# structured graph navigation the load-bearing path — and whether that
-# *substitutes* for the grep/read fan-out (the LocAgent token-savings thesis)
-# instead of being additive overhead. Compare vs the grep/read agent (cost_grep)
-# and the search-only composer.
-sweep_id: cost_locagent
-model: vertex_ai/claude-haiku-4-5
-vertex_location: us-east5
-vertex_project: null
-reps: 3
-max_turns: 16
-temperature: 0.0
-max_tokens: 4096
-topk: 50
-num_retries: 8
-dataset: fishmingyu/codeminer-base-dataset
-split: test
-prebuilt_dir: /mnt/data/codeminer
-embedding_model: Qwen/Qwen3-Embedding-0.6B
-embedding_dimension: 1024
-metrics_k: [1, 3, 5, 10]
-gt_simplified_symbols: true
-include_default_tools: true
-# Graph-primary: read code, but NO grep — structured tools must navigate.
-default_tool_ids: ["read"]
-system_prompt: |
-  You are a code localization agent. Find the code locations (files and symbols)
-  relevant to the request, then give a concise answer naming them.
-
-  You navigate the repository through a code graph and search — there is NO grep.
-  Your tools:
-  - bm25_search(query) — find entry-point symbols by name/identifier.
-  - find_callers(symbol) / find_callees(symbol) — who calls X / what X calls.
-  - trace(from_symbol, to_symbol) — the call path between two symbols.
-  - read(file_path, offset, limit) — read a file / line range to confirm.
-
-  Workflow — SEARCH -> NAVIGATE -> READ -> ANSWER:
-  1. SEARCH — bm25_search for the symbols named/implied by the request.
-  2. NAVIGATE — follow the call graph from those seeds: find_callers /
-     find_callees to move to adjacent code, trace to reach the site across
-     functions. Re-seed with exact symbol names from results.
-  3. READ — read the most promising candidate(s) to confirm the code to
-     change is really there. If not, navigate further from what you just read.
-  4. ANSWER — only once you have READ the implementing file. End with two lines,
-     repo-relative:
-     Files: path/one.ext, path/two.ext
-     Symbols: path/one.ext:symbol_name, ...
-
-  You have a limited tool budget — converge: once a file looks right, confirm and
-  answer rather than exhaustively expanding.
-instances:
-  - redis__redis-10095
-  - tokio-rs__tokio-4898
-  - facebook__docusaurus-10130
-  - axios__axios-4731
-  - hashicorp__terraform-34814
-  - caddyserver__caddy-5870
-  - pydata__xarray-2905
-  - sympy__sympy-13031
-subsets:
-  LOCAGENT:
-    - bm25_search
-    - find_callers
-    - find_callees
-    - trace
