--- scripts/agent_compile/configs/cost_structured.yaml
@@ -1,48 +1,0 @@
-# SPDX-FileCopyrightText: 2025-2026 CodeMiner Contributors
-#
-# SPDX-License-Identifier: Apache-2.0
-
-# Cost study (#133) — STRUCTURED condition: the default file tools are
-# WITHHELD (include_default_tools: false), so the agent must localize with
-# embedding_search (semantic recall) + the call-graph verbs
-# (find_callers/find_callees/trace). This forces the structured/LSP path the
-# agent otherwise refuses to take, to measure whether it reaches the same files
-# as the FREE (grep/read) agent at far fewer tokens.
-#
-# Run per model (same instances as cost_free.yaml):
-#   python scripts/agent_compile/run_agent_sweep.py --config <this> \
-#       --output-dir results/agent_compile/cost_structured_haiku
-#   ... --model vertex_ai/gemini-2.5-flash --vertex-location us-central1 ...
-
-sweep_id: cost_structured
-model: vertex_ai/claude-haiku-4-5
-vertex_location: us-east5
-vertex_project: null
-reps: 1
-max_turns: 12
-temperature: 0.0
-max_tokens: 4096
-topk: 50
-num_retries: 8
-
-dataset: fishmingyu/codeminer-base-dataset
-split: test
-prebuilt_dir: /mnt/data/codeminer
-embedding_model: Qwen/Qwen3-Embedding-0.6B
-embedding_dimension: 1024
-metrics_k: [1, 3, 5, 10]
-gt_simplified_symbols: true
-
-include_default_tools: false   # withhold file_read / file_search
-
-instances:
-  - babel__babel-15445
-  - vuejs__core-11589
-  - micropython__micropython-13569
-
-subsets:
-  STRUCT:
-    - embedding_search
-    - find_callers
-    - find_callees
-    - trace
