--- scripts/agent_compile/configs/phase0.yaml
@@ -1,78 +1,0 @@
-# SPDX-FileCopyrightText: 2025-2026 CodeMiner Contributors
-#
-# SPDX-License-Identifier: Apache-2.0
-
-# Phase 0 kill-switch sweep configuration (issue #133).
-#
-# Run order:
-#   1. python scripts/agent_compile/partition.py \
-#        --instances data/swebench_multilingual_candidates.json \
-#        --seed 20260515 \
-#        --output data/agent_compile/partition.json
-#   2. python scripts/agent_compile/run_sweep.py \
-#        --config scripts/agent_compile/configs/phase0.yaml \
-#        --partition data/agent_compile/partition.json \
-#        --output-dir results/agent_compile/phase0
-#   3. python scripts/agent_compile/aggregate.py \
-#        --cells-dir results/agent_compile/phase0/cells \
-#        --output-dir results/agent_compile/phase0
-#
-# Kill-switch thresholds (also pinned in docs/agent_compile_design.md):
-#   - A6 mean total tokens / query < 30_000
-#   - A6 files@5 ≥ 0.55
-# Both must clear to close the RFC.
-
-sweep_id: phase0
-instances_source: fit          # partition.json -> fit pool (30 instances)
-reps: 3                        # min-of-3 noise handling (per #131 ritual)
-max_turns: 20                  # locked by ADR (fishmingyu 2026-05-08)
-temperature: 0.0
-max_tokens: 4096
-
-dataset: SWE-bench/SWE-bench_Multilingual
-split: test
-languages:
-  - python
-  - go
-  - rust
-  - cpp
-  - c
-  - typescript
-  - javascript
-
-metrics_k: [1, 3, 5]
-
-# Embedding model + dimension match the eval driver's defaults; only
-# loaded when a subset includes a vector-requiring skill (A6 here).
-embedding_model: nomic-ai/CodeRankEmbed
-embedding_dimension: 768
-
-# Cache directories — adjust to taste; leaving null uses defaults.
-index_cache_dir: null
-repo_cache_dir: null
-eval_instances_path: null    # set to gt_locate JSON for symbol-level metrics
-
-# Phase 0 only compares the two endpoints: minimal sparse vs full registry.
-# Phase 2 expands subsets to A1..A5.
-subsets:
-  A0:
-    - bm25_search
-  A6:
-    - bm25_search
-    - embedding_search
-    - hybrid_search
-    - codeminer_context
-    - llm_rerank
-    - find_callers
-    - find_callees
-    - trace
-    - code_to_query
-
-# Phase 0 uses one frontier model — kill-switch decision should not be
-# polluted by inter-model variance. The multi-tier model matrix is described
-# in docs/agent_compile_design.md (no separate config file).
-models:
-  - claude-sonnet-4-6
-
-# Per-environment knobs; merged into the subprocess environment.
-extra_env: {}
