--- scripts/agent_compile/configs/sample.yaml
@@ -1,94 +1,0 @@
-# SPDX-FileCopyrightText: 2025-2026 CodeMiner Contributors
-#
-# SPDX-License-Identifier: Apache-2.0
-
-# Agent-compile SAMPLE sweep (issue #133, Phase 2 dry-run).
-#
-# Runs the full A0..A6 subset matrix over a small multi-language sample of
-# codeminer_base, reusing the offline-built per-instance indexes under
-# prebuilt_dir (see scripts/agent_compile/prebuilt.py). This is the
-# "basic combination test on a few sample instances" deliverable — not the
-# full 30-instance Phase-2 fit pool.
-#
-# Run:
-#   python scripts/agent_compile/run_agent_sweep.py \
-#       --config scripts/agent_compile/configs/sample.yaml \
-#       --output-dir results/agent_compile/sample
-#   python scripts/agent_compile/aggregate_phase2.py \
-#       --cells-dir results/agent_compile/sample/cells \
-#       --output-dir results/agent_compile/sample
-
-sweep_id: sample
-
-# Vertex Anthropic Haiku 4.5 (publisher model lives in us-east5).
-model: vertex_ai/claude-haiku-4-5
-vertex_location: us-east5
-vertex_project: null          # falls back to gcloud ADC default project
-
-reps: 2                       # min-of-N cost, mean-of-N accuracy (per #131 ritual)
-max_turns: 10                 # tightened from 20: the default-tool agent over-
-                              # explored and hit the 20-turn cap 50-100% of the
-                              # time. 10 forces convergence + ~halves cost.
-temperature: 0.0
-max_tokens: 4096
-topk: 50
-
-dataset: fishmingyu/codeminer-base-dataset
-split: test
-prebuilt_dir: /mnt/data/codeminer
-
-# Qwen3-0.6B is pre-cached locally and its FAISS + documents are prebuilt for
-# every instance; nomic/CodeRankEmbed needs trust_remote_code, so we pin Qwen3.
-embedding_model: Qwen/Qwen3-Embedding-0.6B
-embedding_dimension: 1024
-
-metrics_k: [1, 3, 5, 10]
-gt_simplified_symbols: true
-
-# Sample: 5 instances, 4 languages, both Python scenarios (trace / no-trace).
-instances:
-  - astropy__astropy-12907        # Python, no stacktrace
-  - scikit-learn__scikit-learn-13142   # Python, with stacktrace
-  - caddyserver__caddy-5870       # Go, no stacktrace
-  - axios__axios-4731             # TypeScript/JavaScript, no stacktrace
-  - astral-sh__ruff-15309         # Rust, no stacktrace
-
-# A0..A6 sweep ladder over the PRUNED skill set (#133). The always-on default
-# tools (read/grep/glob/bash) are infrastructure, not sweep variables, and are
-# implicit in the agent loop. (The graph leg is the call-graph verbs
-# find_callers/find_callees/trace — the bulk graph_expand skill was removed as
-# it never earned its keep in the agent loop.)
-subsets:
-  A0:
-    - bm25_search
-  A1:
-    - embedding_search
-  A2:
-    - bm25_search
-    - embedding_search
-  A3:
-    - bm25_search
-    - find_callers
-  A4:
-    - bm25_search
-    - embedding_search
-    - find_callers
-    - find_callees
-    - trace
-  A5:
-    - bm25_search
-    - embedding_search
-    - hybrid_search
-    - find_callers
-    - find_callees
-    - trace
-  A6:
-    - bm25_search
-    - embedding_search
-    - hybrid_search
-    - codeminer_context
-    - llm_rerank
-    - find_callers
-    - find_callees
-    - trace
-    - code_to_query
