+++ scripts/analysis/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2025-2026 CodeMiner Contributors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Generic analysis tools over sweep outputs (harness-agnostic)."""
